+++ Traducteur/Mots/Famille.hpp
@@ -28,7 +28,7 @@ class Famille
     public :
     
     Famille();
-    int recuperer_nombre_de_sens_sortie() const { return __sens_sortie.size(); };
+    size_t recuperer_nombre_de_sens_sortie() const { return __sens_sortie.size(); };
     Mot & recuperer_sens_sortie(int numero_sens) { return __sens_sortie[numero_sens]; };
     void definir_les_champs_lexicaux_des_mots(std::vector <std::string> champs_lexicaux);
     void ajouter_sens_sortie(std::string mot);
