+++ Traducteur/Mots/Groupe.hpp
@@ -22,7 +22,7 @@ class Groupe
     
     Groupe(std::string mot_source);
     std::string recuperer_mot_source() const { return _mot_source; };
-    int recuperer_nombre_de_familles() const { return __familles.size(); };
+    size_t recuperer_nombre_de_familles() const { return __familles.size(); };
     Famille & recuperer_famille(int numero_famille) { return __familles[numero_famille]; };
     void ajouter_famille(Famille famille);
